pyproject.toml
[project]
name = "league-data-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "redis>=7.4.0",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
    "pandas>=3.0.1",
    "kafka-python>=2.3.0",
    "pyspark>=4.1.1",
]
